refactor (blocks): show hidden blocks in plugin settings
